move src/db.py -> src/db/connection.py
