+++ src/db/connection.py
@@ -17,7 +17,6 @@ database = Database(settings.DATABASE_URL)
 
 async def db_connection() -> AsyncGenerator:
     """Зависимость для указания в контроллерах Fastapi.
-
     :yields: Connection
     """
     yield database
