+++ src/repositories/auth.py
@@ -10,7 +10,7 @@ from fastapi import Depends
 from pydantic import BaseModel
 from pypika import Query, Table
 
-from db import db_connection
+from db.connection import db_connection
 from exceptions import UserNotFoundError
